Refactoring: move CommandText and CommandOption to base module
